Term 1 Seminar 5 Strings Update

+++ term1/seminar04_types/classroom_tasks/code/1float/24problem_round_solution.c
@@ -3,9 +3,9 @@
 int nearest_integer(float x)
 {
     if (x > 0)
-        return x + 0.5;
+        return (int)(x + 0.5);
     else
-        return x - 0.5;
+        return (int)(x - 0.5);
 }
